+++ src/modules/admin/components/ceMatchRules/CeMatchRuleForm.tsx
@@ -1,16 +1,11 @@
-import CodeIcon from '@mui/icons-material/Code';
-import { Alert, Button, Stack, Typography } from '@mui/material';
+import { Alert, Button, Stack } from '@mui/material';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 
-import {
-  CeMatchExpressionMode,
-  CeMatchRuleFormValues,
-  newDraftClause,
-} from './ceMatchRuleFormUtil';
+import CeMatchExpressionModeSwitch from './CeMatchExpressionModeSwitch';
+import { CeMatchRuleFormValues, newDraftClause } from './ceMatchRuleFormUtil';
 import FreeTextExpressionEditor from './FreeTextExpressionEditor';
 import StructuredExpressionBuilder from './StructuredExpressionBuilder';
-import ConfirmationDialog from '@/components/elements/ConfirmationDialog';
 import LoadingButton from '@/components/elements/LoadingButton';
 import TitleCard from '@/components/elements/TitleCard';
 import ApolloErrorAlert from '@/modules/errors/components/ApolloErrorAlert';
@@ -38,28 +33,6 @@ const defaultCeMatchRuleFormValues = (): CeMatchRuleFormValues => ({
   },
   freeTextExpression: '',
 });
-
-const hasValue = (value?: string | null) => !!value?.trim();
-
-const hasExpressionProgress = ({
-  mode,
-  structuredExpression,
-  freeTextExpression,
-}: CeMatchRuleFormValues) => {
-  if (mode === 'freeText') return hasValue(freeTextExpression);
-
-  return (
-    structuredExpression.clauses.length > 1 ||
-    structuredExpression.operator !== CeMatchRuleBooleanOperator.And ||
-    structuredExpression.clauses.some(
-      ({ source, customAssessmentFormIdentifier, field, value }) =>
-        hasValue(source) ||
-        hasValue(customAssessmentFormIdentifier) ||
-        hasValue(field) ||
-        hasValue(value)
-    )
-  );
-};
 
 const buildCeMatchRuleInput = ({
   name,
@@ -108,7 +81,6 @@ const CeMatchRuleForm = () => {
 
   const [errorState, setErrorState] = useState<ErrorState>(emptyErrorState);
   const [saved, setSaved] = useState(false);
-  const [pendingMode, setPendingMode] = useState<CeMatchExpressionMode>();
   const mode = watch('mode');
 
   const resetForm = () => {
@@ -146,36 +118,8 @@ const CeMatchRuleForm = () => {
     });
   };
 
-  const switchEditorMode = (nextMode: CeMatchExpressionMode) => {
-    setValue('mode', nextMode, { shouldDirty: true });
-    setPendingMode(undefined);
-  };
-
-  const handleSwitchEditorMode = () => {
-    const nextMode = mode === 'structured' ? 'freeText' : 'structured';
-    if (hasExpressionProgress(getValues())) {
-      setPendingMode(nextMode);
-      return;
-    }
-
-    switchEditorMode(nextMode);
-  };
-
   return (
     <Stack gap={2}>
-      <ConfirmationDialog
-        id='confirmSwitchExpressionEditor'
-        open={!!pendingMode}
-        title='Switch Expression Editor?'
-        loading={false}
-        confirmText='Switch Editor'
-        onConfirm={() => pendingMode && switchEditorMode(pendingMode)}
-        onCancel={() => setPendingMode(undefined)}
-      >
-        <Typography variant='body2'>
-          Switching editors will clear your current progress.
-        </Typography>
-      </ConfirmationDialog>
       <TitleCard title='Rule Details' headerComponent='h2' padded>
         <Stack gap={2}>
           {saved && <Alert severity='success'>Match rule saved.</Alert>}
@@ -199,16 +143,11 @@ const CeMatchRuleForm = () => {
         headerComponent='h2'
         padded
         actions={
-          <Button
-            size='small'
-            variant='text'
-            startIcon={<CodeIcon />}
-            onClick={handleSwitchEditorMode}
-          >
-            {mode === 'structured'
-              ? 'Switch to Advanced Expression Editor'
-              : 'Switch to Structured Editor'}
-          </Button>
+          <CeMatchExpressionModeSwitch
+            mode={mode}
+            getValues={getValues}
+            setValue={setValue}
+          />
         }
       >
         <Stack gap={2}>
